src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-openai==0.3.33
python-multipart==0.0.6